requirements.txt
pillow~=10.2.0
pygame~=2.0.0